final submission q1 + q2
