+++ hw2/iar1000/q1/Line.h
@@ -1,6 +1,7 @@
 #include <stddef.h>
 #include <string>
 #include <vector>
+#include <limits>
 
 class Line
   {
